Fix types

+++ src/bf.cpp
@@ -1,7 +1,7 @@
 #include <blowfish.h>
 #include "bf.h"
 
-size_t BF::encrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, size_t key_size)
+size_t BF::encrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, int key_size)
 {
     Blowfish blowfish;
     blowfish.SetKey(key, key_size);
@@ -10,7 +10,7 @@ size_t BF::encrypt(const std::vector<unsigned char> &input_data, std::vector<uns
     return output_data.size();
 }
 
-size_t BF::decrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, size_t key_size)
+size_t BF::decrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, int key_size)
 {
     Blowfish blowfish;
     blowfish.SetKey(key, key_size);
