+++ src/bf.h
@@ -7,8 +7,8 @@
 
 namespace BF
 {
-    size_t encrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, size_t key_size);
-    size_t decrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, size_t key_size);
+    size_t encrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, int key_size);
+    size_t decrypt(const std::vector<unsigned char> &input_data, std::vector<unsigned char> &output_data, const unsigned char *key, int key_size);
 }
 
 #endif // BF_H
